change PDO Error-Mode for PHP 8
Starting with PHP 8.0.0 the PDO Error-Mode is set to PDO::ERRMODE_EXCEPTION as default (Prior it was PDO::ERRMODE_SILENT).
This leads to an uncaught PDOexception in db.php function tableExists after clean install with an empty database. Of cause the resulting admin-page is empty with a 500 Internal server fault.

I fixed this for me by changing the error-mode to the old default value after connect.